requirements.txt
autopep8==2.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==22.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
